requirements.txt
amqp==5.1.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.52.0
django-countries==7.5.1
django-filter==22.1
django-haystack==3.1.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-extra-fields==3.4.1
drf-haystack==1.8.11
idna==3.4
kombu==5.2.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
phonenumbers==8.13.6
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
serpy==0.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
